+++ src/test/java/chess/dao/ChessDAOTest.java
@@ -1,5 +1,6 @@
 package chess.dao;
 
+import chess.entity.Chess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
